+++ libs/cdk/render-strategies/src/lib/onStrategy.ts
@@ -1,3 +1,4 @@
+import { NgZone } from '@angular/core';
 import { RxCoalescingOptions } from '@rx-angular/cdk/coalescing';
 import { switchMap, take } from 'rxjs/operators';
 import { Observable, of, throwError } from 'rxjs';
@@ -19,22 +20,24 @@ export function onStrategy<T>(
     work: RxRenderWork,
     options: RxCoalescingOptions
   ) => void,
-  options: RxCoalescingOptions = {}
+  options: RxCoalescingOptions & { ngZone?: NgZone } = {}
 ): Observable<T> {
   let error: Error;
-  return new Observable<T>(subscriber => {
+  return new Observable<T>((subscriber) => {
     subscriber.next(value);
   }).pipe(
-    strategy.behavior(() => {
-      try {
-        workFactory(value, strategy.work, options);
-      } catch (e) {
-        error = e;
-      }
-    }, options.scope || {}),
-    switchMap(() =>
-      error ? throwError([error, value]) : of(value)
-    ),
+    strategy.behavior({
+      work: () => {
+        try {
+          workFactory(value, strategy.work, options);
+        } catch (e) {
+          error = e;
+        }
+      },
+      scope: (options.scope as Record<string, unknown>) || {},
+      ngZone: options.ngZone,
+    }),
+    switchMap(() => (error ? throwError([error, value]) : of(value))),
     take(1)
   );
 }